add temporary simulator
